requirements.txt
tabulate==0.8.6
pyyaml==5.3
